chore: split hooks into files
